+++ pkgs/jnigen/java/src/main/java/com/github/dart_lang/jnigen/apisummarizer/disasm/AsmMethodSignatureVisitor.java
@@ -25,6 +25,7 @@ public class AsmMethodSignatureVisitor extends SignatureVisitor {
     var typeParam = new TypeParam();
     typeParam.name = name;
     method.typeParams.add(typeParam);
+    super.visitFormalTypeParameter(name);
   }
 
   @Override